serving/requirements.txt
flask==2.2.3
flask-cors==3.0.10
flask-jwt-extended==4.4.4
flask-limiter==3.3.0
gunicorn==20.1.0
pillow==9.4.0
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.28.2
torch==2.0.0
transformers==4.27.0
numpy==1.24.2
pandas==1.5.3
networkx==3.0
matplotlib==3.7.1
scikit-learn==1.2.2 